refactor: move units code out of helpers.py to units.helpers.py for now
