+++ pyaerocom/aeroval/helpers.py
@@ -7,8 +7,8 @@ from pydantic import BaseModel
 from pyaerocom import const
 from pyaerocom.aeroval.modelentry import ModelEntry
 from pyaerocom.griddeddata import GriddedData
+from pyaerocom.units.helpers import get_highest_resolution
 from pyaerocom.helpers import (
-    get_highest_resolution,
     get_max_period_range,
     make_dummy_cube,
     start_stop,
