+++ pyaerocom/helpers.py
@@ -6,8 +6,7 @@ from __future__ import annotations
 
 import logging
 import math as ma
-from collections import Counter
-from datetime import MINYEAR, date, datetime
+from datetime import MINYEAR, datetime
 
 import iris
 import iris.analysis
@@ -19,21 +18,18 @@ import xarray as xr
 from cf_units import Unit
 
 from pyaerocom import const
-from pyaerocom._warnings import ignore_warnings
 from pyaerocom.exceptions import (
     DataCoverageError,
     DataDimensionError,
     LongitudeConstraintError,
     MetaDataError,
     ResamplingError,
-    TemporalResolutionError,
     VariableDefinitionError,
 )
 from pyaerocom.units.time_config import (
     GREGORIAN_BASE,
     PANDAS_RESAMPLE_OFFSETS,
     TS_TYPE_DATETIME_CONV,
-    TS_TYPE_SECS,
     TS_TYPE_TO_PANDAS_FREQ,
     TS_TYPE_TO_FREQ_NAME,
     TS_TYPE_TO_NUMPY_FREQ,
@@ -46,6 +42,8 @@ from pyaerocom.units.time_config import (
 )
 from pyaerocom.units.tstype import TsType
 from pyaerocom.variable_helpers import get_variable
+from pyaerocom.units.helpers import is_year, to_pandas_timestamp
+
 
 logger = logging.getLogger(__name__)
 
@@ -414,164 +412,6 @@ def copy_coords_cube(to_cube, from_cube, inplace=True):
     return to_cube
 
 
-def infer_time_resolution(time_stamps, dt_tol_percent=5, minfrac_most_common=0.8):
-    """Infer time resolution based on input time-stamps
-
-    Calculates time difference *dt* between consecutive timestamps provided via
-    input array or list. Then it counts the most common *dt* (e.g. 86400 s for
-    daily). Before inferring the frequency it then checks all other *dts*
-    occurring in the input array to see if they are within a certain interval
-    around the most common one (e.g. +/- 5% as default, via arg
-    `dt_tol_percent`), that is, 86390 would be included if most common dt is
-    86400 s but not 80000s. Then it checks if the number of *dts* that
-    are within that tolerance level around the most common *dt* exceed a
-    certain fraction (arg `minfrac_most_common`) of the total number of *dts*
-    that occur in the input array (default is 80%). If that is the case, the
-    most common frequency is attempted to be derived using
-    :func:`TsType.from_total_seconds` based on the most common *dt* (in this
-    example that would be *daily*).
-
-
-    Parameters
-    ----------
-    time_stamps : pandas.DatetimeIndex, or similar
-        list of time stamps
-    dt_tol_percent : int
-        tolerance in percent of accepted range of time diffs with respect to
-        most common time difference.
-    minfrac_most_common : float
-        minimum required fraction of time diffs that have to be equal to, or
-        within tolerance range, the most common time difference.
-
-
-    Raises
-    ------
-    TemporalResolutionError
-        if frequency cannot be derived.
-
-    Returns
-    -------
-    str
-        inferred frequency
-    """
-    from pyaerocom import TsType
-
-    if not isinstance(time_stamps, pd.DatetimeIndex):
-        time_stamps = pd.DatetimeIndex(time_stamps)
-    vals = time_stamps.values
-
-    dts = (vals[1:] - vals[:-1]).astype("timedelta64[s]").astype(int)
-
-    if np.min(dts) < 0:
-        raise TemporalResolutionError("Nasa Ames file contains neg. meas periods...")
-
-    counts = Counter(dts).most_common()
-    most_common_dt, most_common_num = counts[0]
-    num_within_tol = most_common_num
-    lower = most_common_dt * (100 - dt_tol_percent) / 100
-    upper = most_common_dt * (100 + dt_tol_percent) / 100
-    for dt, num in counts[1:]:
-        if lower <= dt <= upper:
-            num_within_tol += num
-    frac_ok = num_within_tol / len(dts)
-    if not frac_ok > minfrac_most_common:
-        raise TemporalResolutionError("Failed to infer ts_type")
-    tst = TsType.from_total_seconds(most_common_dt)
-    return str(tst)
-
-
-def seconds_in_periods(timestamps, ts_type):
-    """
-    Calculates the number of seconds for each period in timestamps.
-
-    Parameters
-    ----------
-    timestamps : numpy.datetime64 or numpy.ndarray
-        Either a single datetime or an array of datetimes.
-    ts_type : str
-        Frequency of timestamps.
-
-    Returns
-    -------
-    np.array :
-        Array with same length as timestamps containing number of seconds for
-        each period.
-    """
-
-    ts_type = TsType(ts_type)
-    if isinstance(timestamps, np.datetime64):
-        timestamps = np.array([timestamps])
-    if isinstance(timestamps, np.ndarray):
-        timestamps = [to_pandas_timestamp(timestamp) for timestamp in timestamps]
-    # From here on timestamps should be a numpy array containing pandas Timestamps
-
-    seconds_in_day = 86400
-    if ts_type >= TsType("monthly"):
-        if ts_type == TsType("monthly"):
-            days_in_months = np.array([timestamp.days_in_month for timestamp in timestamps])
-            seconds = days_in_months * seconds_in_day
-            return seconds
-        if ts_type == TsType("daily"):
-            return seconds_in_day * np.ones_like(timestamps)
-        raise NotImplementedError("Only yearly, monthly and daily frequencies implemented.")
-    elif ts_type == TsType("yearly"):
-        days_in_year = []
-        for ts in timestamps:
-            if ts.year % 4 == 0:
-                days_in_year.append(366)  #  Leap year
-            else:
-                days_in_year.append(365)
-        seconds = np.array(days_in_year) * seconds_in_day
-        return seconds
-    raise TemporalResolutionError(f"Unknown TsType: {ts_type}")
-
-
-def get_tot_number_of_seconds(ts_type, dtime=None):
-    """Get total no. of seconds for a given frequency
-
-    ToDo
-    ----
-    This method needs revision and can be solved simpler probably
-
-    Parameters
-    ----------
-    ts_type : str or TsType
-        frequency for which number of seconds is supposed to be retrieved
-    dtime : TYPE, optional
-        DESCRIPTION. The default is None.
-
-    Raises
-    ------
-    AttributeError
-        DESCRIPTION.
-
-    Returns
-    -------
-    TYPE
-        DESCRIPTION.
-
-    """
-
-    ts_tpe = TsType(ts_type)
-
-    if ts_tpe >= TsType("monthly"):
-        if dtime is None:
-            raise AttributeError(
-                "For frequncies larger than or eq. monthly you"
-                + " need to provide dtime in order to compute the number of second."
-            )
-        if not ts_type == "monthly":
-            raise NotImplementedError("Can only handle monthly so far...")
-
-        # find seconds from dtime
-        # TODO generalize this
-        days_in_month = dtime.dt.daysinmonth
-
-        return days_in_month * 24 * 60 * 60
-    else:
-        return TS_TYPE_SECS[ts_type]
-
-
 def get_standard_name(var_name):
     """Converts AeroCom variable name to CF standard name
 
@@ -591,120 +431,6 @@ def get_standard_name(var_name):
     from pyaerocom import const
 
     return const.VARS[var_name].standard_name
-
-
-def get_standard_unit(var_name):
-    """Gets standard unit of AeroCom variable
-
-    Also handles alias names for variables, etc. or strings corresponding to
-    older conventions (e.g. names containing 3D).
-
-    Parameters
-    ----------
-    var_name : str
-        AeroCom variable name
-
-    Returns
-    -------
-    str
-        corresponding standard unit
-    """
-    from pyaerocom import const
-
-    return const.VARS[var_name].units
-
-
-def get_lowest_resolution(ts_type, *ts_types):
-    """Get the lowest resolution from several ts_type codes
-
-    Parameters
-    ----------
-    ts_type : str
-        first ts_type
-    *ts_types
-        one or more additional ts_type codes
-
-    Returns
-    -------
-    str
-        the ts_type that corresponds to the lowest resolution
-
-    Raises
-    ------
-    ValueError
-        if one of the input ts_type codes is not supported
-    """
-    # all_ts_types = const.GRID_IO.TS_TYPES
-    from pyaerocom.units.tstype import TsType
-
-    lowest = TsType(ts_type)
-    for freq in ts_types:
-        _temp = TsType(freq)
-        if _temp < lowest:
-            lowest = _temp
-    return lowest.val
-
-
-def sort_ts_types(ts_types):
-    """Sort a list of ts_types
-
-    Parameters
-    ----------
-    ts_types : list
-        list of strings (or instance of :class:`TsType`) to be sorted
-
-    Returns
-    -------
-    list
-        list of strings with sorted frequencies
-
-    Raises
-    ------
-    TemporalResolutionError
-        if one of the input ts_types is not supported
-    """
-    freqs_sorted = []
-    for ts_type in ts_types:
-        if isinstance(ts_type, str):
-            ts_type = TsType(ts_type)
-        if len(freqs_sorted) == 0:
-            freqs_sorted.append(ts_type)
-        else:
-            insert = False
-            for i, tt in enumerate(freqs_sorted):
-                if tt < ts_type:
-                    insert = True
-                    break
-            if insert:
-                freqs_sorted.insert(i, ts_type)
-            else:
-                freqs_sorted.append(ts_type)
-    return [str(tt) for tt in freqs_sorted]
-
-
-def get_highest_resolution(ts_type, *ts_types):
-    """Get the highest resolution from several ts_type codes
-
-    Parameters
-    ----------
-    ts_type : str
-        first ts_type
-    *ts_types
-        one or more additional ts_type codes
-
-    Returns
-    -------
-    str
-        the ts_type that corresponds to the highest resolution
-
-    Raises
-    ------
-    ValueError
-        if one of the input ts_type codes is not supported
-    """
-    lst = [ts_type]
-    lst.extend(ts_types)
-    return sort_ts_types(lst)[0]
 
 
 def isnumeric(val):
@@ -1179,8 +905,6 @@ def resample_time_dataarray(arr, freq, how=None, min_num_obs=None):
     elif "time" not in arr.dims:
         raise DataDimensionError("Cannot resample time: input DataArray has no time dimension")
 
-    from pyaerocom.units.tstype import TsType
-
     to = TsType(freq)
     pd_freq = to.to_pandas_freq()
     invalid = None
@@ -1265,80 +989,6 @@ def str_to_iris(key, **kwargs):
     if callable(val):
         return val(**kwargs)
     return val
-
-
-@ignore_warnings(UserWarning, r"Parsing .* in DD/MM/YYYY format")
-def to_pandas_timestamp(value):
-    """Convert input to instance of :class:`pandas.Timestamp`
-
-    Parameters
-    ----------
-    value
-        input value that is supposed to be converted to time stamp
-
-    Returns
-    --------
-    pandas.Timestamp
-    """
-    if isinstance(value, np.str_):
-        value = str(value)
-    if isinstance(value, pd.Timestamp):
-        return value
-    elif isinstance(value, str | np.datetime64 | datetime | date):
-        return pd.Timestamp(value)
-    else:
-        try:
-            numval = int(value)
-            if not 0 <= numval <= 10000:
-                raise ValueError("Could not infer valid year from numerical time input")
-            return pd.Timestamp(str(numval))
-        except Exception as e:
-            raise ValueError(f"Failed to convert {value} to Timestamp: {repr(e)}")
-
-
-def to_datetime64(value):
-    """Convert input value to numpy.datetime64
-
-    Parameters
-    ----------
-    value
-        input value that is supposed to be converted, needs to be either str,
-        datetime.datetime, pandas.Timestamp or an integer specifying the
-        desired year.
-
-    Returns
-    -------
-    datetime64
-        input timestamp converted to datetime64
-    """
-    if isinstance(value, np.datetime64):
-        return value
-    else:
-        try:
-            return to_pandas_timestamp(value).to_datetime64()
-        except Exception as e:
-            raise ValueError(f"Failed to convert {value} to datetime64 objectError: {repr(e)}")
-
-
-def is_year(val):
-    """Check if input is / may be year
-
-    Parameters
-    ----------
-    val
-        input that is supposed to be checked
-
-    Returns
-    -------
-    bool
-        True if input is a number between -2000 and 10000, else False
-    """
-    try:
-        if -2000 < int(val) < 10000:
-            return True
-        raise Exception
-    except Exception:
-        return False
 
 
 def _check_climatology_timestamp(t):
